[Refactor] Flatten nested folder for all protobuf files
