+++ BUILD.bazel
@@ -1134,15 +1134,6 @@ genrule(
         # https://github.com/ray-project/ray/issues/31358
         sed -i -E 's/from opencensus.proto.metrics.v1 import/from . import/' "$${files[@]}"
         sed -i -E 's/from opencensus.proto.resource.v1 import/from . import/' "$${files[@]}"
-        # TODO(rickyx): Remove this experimental path rewrite once we move autoscaler proto out
-        # of experimental.
-        autoscale_files=(`ls python/ray/core/generated/instance_manager_pb2*.py`)
-        sed -i -E 's/from ..experimental/from ./' "$${autoscale_files[@]}"
-        # Help the generated export api files to have the correct module
-        export_api_files=(`ls python/ray/core/generated/export*_pb2*.py`)
-        sed -i -E 's/from ..export_api/from ./' "$${export_api_files[@]}"
-        events_files=(`ls python/ray/core/generated/events*_pb2*.py`)
-        sed -i -E 's/from ..events/from ./' "$${events_files[@]}"
         echo "$${PWD}" > $@
     """,
     local = 1,
