+++ src/ray/common/ray_config_def.h
@@ -942,7 +942,7 @@ RAY_CONFIG(bool, enable_export_api_write, false)
 // types to write export API events for. This configuration is only used if
 // RAY_enable_export_api_write is not enabled. Full list of valid
 // resource types in ExportEvent.SourceType enum in
-// src/ray/protobuf/export_api/export_event.proto
+// src/ray/protobuf/export_event.proto
 // Example config: `export RAY_enable_export_api_write_config='EXPORT_ACTOR,EXPORT_TASK'`
 RAY_CONFIG(std::vector<std::string>, enable_export_api_write_config, {})
