+++ src/ray/protobuf/BUILD
@@ -35,7 +35,7 @@ proto_library(
 
 proto_library(
     name = "instance_manager_proto",
-    srcs = ["experimental/instance_manager.proto"],
+    srcs = ["instance_manager.proto"],
 )
 
 python_grpc_compile(
@@ -248,7 +248,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_event_proto",
-    srcs = ["export_api/export_event.proto"],
+    srcs = ["export_event.proto"],
     deps = [
         ":export_actor_event_proto",
         ":export_dataset_metadata_proto",
@@ -272,7 +272,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_task_event_proto",
-    srcs = ["export_api/export_task_event.proto"],
+    srcs = ["export_task_event.proto"],
     deps = [
         ":common_proto",
         ":export_runtime_env_proto",
@@ -291,7 +291,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_runtime_env_proto",
-    srcs = ["export_api/export_runtime_env.proto"],
+    srcs = ["export_runtime_env.proto"],
 )
 
 cc_proto_library(
@@ -306,7 +306,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_node_event_proto",
-    srcs = ["export_api/export_node_data.proto"],
+    srcs = ["export_node_data.proto"],
 )
 
 cc_proto_library(
@@ -321,7 +321,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_actor_event_proto",
-    srcs = ["export_api/export_actor_data.proto"],
+    srcs = ["export_actor_data.proto"],
     deps = [":common_proto"],
 )
 
@@ -337,7 +337,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_driver_job_event_proto",
-    srcs = ["export_api/export_driver_job_event.proto"],
+    srcs = ["export_driver_job_event.proto"],
     deps = [
         ":common_proto",
         ":export_runtime_env_proto",
@@ -356,7 +356,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_submission_job_event_proto",
-    srcs = ["export_api/export_submission_job_event.proto"],
+    srcs = ["export_submission_job_event.proto"],
 )
 
 cc_proto_library(
@@ -371,7 +371,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_train_state_proto",
-    srcs = ["export_api/export_train_state.proto"],
+    srcs = ["export_train_state.proto"],
 )
 
 cc_proto_library(
@@ -386,7 +386,7 @@ python_grpc_compile(
 
 proto_library(
     name = "export_dataset_metadata_proto",
-    srcs = ["export_api/export_dataset_metadata.proto"],
+    srcs = ["export_dataset_metadata.proto"],
 )
 
 cc_proto_library(
@@ -510,7 +510,7 @@ cc_proto_library(
 
 proto_library(
     name = "events_actor_task_definition_event_proto",
-    srcs = ["events/events_actor_task_definition_event.proto"],
+    srcs = ["events_actor_task_definition_event.proto"],
     deps = [
         ":common_proto",
         ":runtime_env_common_proto",
@@ -529,7 +529,7 @@ python_grpc_compile(
 
 proto_library(
     name = "events_actor_task_execution_event_proto",
-    srcs = ["events/events_actor_task_execution_event.proto"],
+    srcs = ["events_actor_task_execution_event.proto"],
     deps = [
         ":common_proto",
         "@com_google_protobuf//:timestamp_proto",
@@ -548,7 +548,7 @@ python_grpc_compile(
 
 proto_library(
     name = "events_task_definition_event_proto",
-    srcs = ["events/events_task_definition_event.proto"],
+    srcs = ["events_task_definition_event.proto"],
     deps = [
         ":common_proto",
         ":runtime_env_common_proto",
@@ -567,7 +567,7 @@ python_grpc_compile(
 
 proto_library(
     name = "events_task_execution_event_proto",
-    srcs = ["events/events_task_execution_event.proto"],
+    srcs = ["events_task_execution_event.proto"],
     deps = [
         ":common_proto",
         "@com_google_protobuf//:timestamp_proto",
@@ -586,7 +586,7 @@ python_grpc_compile(
 
 proto_library(
     name = "events_base_event_proto",
-    srcs = ["events/events_base_event.proto"],
+    srcs = ["events_base_event.proto"],
     deps = [
         ":events_actor_task_definition_event_proto",
         ":events_actor_task_execution_event_proto",
@@ -608,7 +608,7 @@ python_grpc_compile(
 
 proto_library(
     name = "events_event_aggregator_service_proto",
-    srcs = ["events/events_event_aggregator_service.proto"],
+    srcs = ["events_event_aggregator_service.proto"],
     deps = [
         ":common_proto",
         ":events_base_event_proto",
