+++ src/ray/protobuf/events_base_event.proto
@@ -17,10 +17,10 @@ syntax = "proto3";
 package ray.rpc.events;
 
 import "google/protobuf/timestamp.proto";
-import "src/ray/protobuf/events/events_actor_task_definition_event.proto";
-import "src/ray/protobuf/events/events_actor_task_execution_event.proto";
-import "src/ray/protobuf/events/events_task_definition_event.proto";
-import "src/ray/protobuf/events/events_task_execution_event.proto";
+import "src/ray/protobuf/events_actor_task_definition_event.proto";
+import "src/ray/protobuf/events_actor_task_execution_event.proto";
+import "src/ray/protobuf/events_task_definition_event.proto";
+import "src/ray/protobuf/events_task_execution_event.proto";
 
 // This is the base message for all ray events.
 message RayEvent {
