+++ src/ray/protobuf/events_event_aggregator_service.proto
@@ -15,7 +15,7 @@
 syntax = "proto3";
 
 import "src/ray/protobuf/common.proto";
-import "src/ray/protobuf/events/events_base_event.proto";
+import "src/ray/protobuf/events_base_event.proto";
 
 package ray.rpc.events;
